requirements.txt
absl-py==0.15.0
alabaster==0.7.12
aniso8601==9.0.1
appdirs==1.4.4
argh==0.26.2
arrow==0.13.1
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
black==19.10b0
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2020.6.20
charset-normalizer==2.0.12
clang==5.0
click==8.0.3
cycler==0.11.0
dataclasses==0.8
entrypoints==0.3
Flask==2.0.3
Flask-RESTful==0.3.9
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
gunicorn==20.1.0
h5py==3.1.0
importlib-metadata==4.8.3
inflection==0.5.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-vggface==0.6
kiwisolver==1.3.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.0.4
mkl-service==2.3.0
mtcnn==0.1.1
mypy-extensions==0.4.3
numpy==1.19.5
oauthlib==3.2.0
opencv-python==4.5.5.62
opt-einsum==3.3.0
parso==0.7.0
pathspec==0.7.0
pickle5==0.0.12
Pillow==8.4.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.3
pywin32==228
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==0.24.2
scipy==1.5.4
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
tensorflow-gpu==2.6.0
termcolor==1.1.0
threadpoolctl==3.1.0
webencodings==0.5.1
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.12.1
